Use BigDecimal to create EDI

+++ src/main/java/com/pepe/albarapp/service/document/aecoc/IgicTax.java
@@ -1,5 +1,7 @@
 package com.pepe.albarapp.service.document.aecoc;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 
 import lombok.Data;
@@ -13,7 +15,7 @@ public class IgicTax {
 	private AmountWrapper baseAmount;
 	private AmountWrapper taxAmount;
 
-	public IgicTax(double grossTotal, double taxTotal, double percentage) {
+	public IgicTax(BigDecimal grossTotal, BigDecimal taxTotal, BigDecimal percentage) {
 		this.baseAmount = new AmountWrapper(grossTotal);
 		this.taxAmount = new AmountWrapper(taxTotal);
 		this.percentage = String.valueOf(percentage);
